sentiment-analysis-tool/requirements.txt
requests
pandas
numpy
scikit-learn
pyyaml
openai  # or any other library for the large language model API
nltk  # for natural language processing tasks
textblob  # for sentiment analysis tasks
newsapi-python # for fetching news articles
transformers # for pre-trained models
torch # for deep learning tasks
torchvision # for computer vision tasks
torchaudio # for audio tasks
matplotlib # for plotting
